Implement embedded Qt kernel manager.
